weightset: q: fix less
In [2]: c = vcsn.Q.expression('<10>a+<-10>a', 'none').standard()
        c.lightest()
Out[2]: <10>a

The less comparison will compute:
  10 * 1 < -10 * 1
Which will turn operands into unsigned int and return true.

* vcsn/weightset/q.hh: Here.
* tests/python/less-than.py: Add tests here.

+++ vcsn/weightset/q.hh
@@ -156,7 +156,8 @@ namespace vcsn
     /// Whether \a lhs < \a rhs.
     static bool less(const value_t lhs, const value_t rhs)
     {
-      return lhs.num * rhs.den < rhs.num * lhs.den;
+      return lhs.num * static_cast<long>(rhs.den)
+             < rhs.num * static_cast<long>(lhs.den);
     }
 
     static constexpr bool is_commutative() { return true; }
